// frontend/src/module/auth/schema/register.schema.ts
import { z } from "zod"

export const registerSchema = z.object({
  name: z.string().min(2, "Tên phải có ít nhất 2 ký tự"),
  email: z.string().email("Email không hợp lệ"),
  password: z.string().min(6, "Mật khẩu tối thiểu 6 ký tự"),
})

export type RegisterForm = z.infer<typeof registerSchema>